updated selenium_test_runner
- added command line arguments for receiving web driver path
- implemented use of web drive path argument
- updated default test_dir and results_dir paths
- screenshot dir and web driver path variables are written to selenium
cfg file for use by the test manager

+++ selenium_unittest/selenium_test_runner.py
@@ -9,6 +9,15 @@ parser.add_argument('--browser_name',
                     help='Browser to run tests on.',
                     choices=['Chrome'],
                     required=True)
+parser.add_argument('--webdriver_path',
+                    dest='webdriver_path',
+                    help='Path to webdriver executable.',
+                    required=True)
+parser.add_argument('--webdriver_path_type',
+                    dest='webdriver_path_type',
+                    help='Type of path to use when determining webdriver_path.',
+                    choices=['absolute', 'relative'],
+                    default='relative')
 parser.add_argument('--browser_version',
                     dest='browser_version',
                     help='Selenium test browser version.',
@@ -17,7 +26,7 @@ parser.add_argument('--test_dir',
                     dest='test_dir',
                     help='Path to directory containing tests. '
                          '(path type is based on value of --test_dir_path_type which defaults to relative)',
-                    default='~')
+                    default='')
 parser.add_argument('--test_dir_path_type',
                     dest='test_dir_path_type',
                     help='Type of path to use when determining test_dir location.',
@@ -27,7 +36,7 @@ parser.add_argument('--results_dir',
                     dest='results_dir',
                     help='Path to directory to read/write results from/to.'
                          '(path type is based on value of --results_dir_path_type which defaults to relative)',
-                    default='~')
+                    default='results')
 parser.add_argument('--results_dir_path_type',
                     dest='results_dir_path_type',
                     help='Type of path to use when determining results_dir location.',
@@ -52,15 +61,27 @@ parser.add_argument('--base_url',
                     default='https://beta.guidebook.com')
 args = parser.parse_args()
 
+# determine start and results directory paths
+import os
+current_working_directory = os.getcwd()
+start_dir = os.path.expanduser(args.test_dir) if args.test_dir_path_type =='absolute' else os.path.normpath('%s/%s' %(current_working_directory, args.test_dir))
+if not os.path.isdir(start_dir):
+    raise Exception('start_dir "%s" does not exist!' %start_dir)
+webdriver_path = os.path.expanduser(args.webdriver_path) if args.webdriver_path_type =='absolute' else os.path.normpath('%s/%s' %(current_working_directory, args.webdriver_path))
+if not os.path.exists(webdriver_path):
+    raise Exception('webdriver_path "%s" does not exist!' %webdriver_path)
+results_dir = os.path.expanduser(args.results_dir) if args.results_dir_path_type =='absolute' else os.path.normpath('%s/%s' %(current_working_directory, args.results_dir))
+results_file_path = os.path.join(results_dir, args.browser_version, '%s.json' %(args.browser_name))
+result_screenshots_dir = os.path.join(results_dir, 'screenshots')
 # write selenium parameters to file
 import json
 kwargs = {
     'browser_name': args.browser_name,
     'browser_version': args.browser_version,
-    'base_url': args.base_url
+    'webdriver_path': webdriver_path,
+    'base_url': args.base_url,
+    'screenshot_dir': result_screenshots_dir
 }
-import os
-current_working_directory = os.getcwd()
 selenium_config_file_path = os.path.join(current_working_directory, "selenium.cfg")
 json.dump(kwargs, open(selenium_config_file_path,'w'))
 
@@ -76,12 +97,6 @@ def run_tests(start_dir, pattern, top_level_dir):
             for test in suite._tests:
                 test.__dict__.update(kwargs.items())
     return test_modules
-start_dir = os.path.expanduser(args.test_dir) if args.test_dir_path_type =='absolute' else os.path.normpath('%s/%s' %(current_working_directory, args.test_dir))
-if not os.path.isdir(start_dir):
-    raise Exception('start_dir "%s" does not exist!' %start_dir)
-results_dir = os.path.expanduser(args.results_dir) if args.results_dir_path_type =='absolute' else os.path.normpath('%s/%s' %(current_working_directory, args.results_dir))
-results_file_path = os.path.join(results_dir, args.browser_version, '%s.json' %(args.browser_name))
-result_screenshots_dir = os.path.join(results_dir, 'screenshots')
 # delete previous results if show_previous_results not specified
 if not args.show_previous_results and os.path.exists(results_file_path):
     os.remove(results_file_path)
